+++ esp_video/CHANGELOG.md
@@ -8,6 +8,7 @@
   - YUV422 supports two byte orders: YUYV and UYVY, represented by `V4L2_PIX_FMT_YUYV` and `V4L2_PIX_FMT_UYVY` respectively; RGB565 supports two byte orders: RGB565_LE and RGB565_BE, represented by `V4L2_PIX_FMT_RGB565` and `V4L2_PIX_FMT_RGB565X` respectively.
   - `V4L2_PIX_FMT_YUV422P` format is no longer used in the application; replace it with formats `V4L2_PIX_FMT_YUYV` or `V4L2_PIX_FMT_UYVY`.
 - Add support for 2-bit data I/O mode on SPI video device when using parallel I/O interface
+- Add set/get DQBUF timeout value
 
 ## 1.4.1